multinano calibration and start/stop eyeserver

+++ cobe/app.py
@@ -72,14 +72,32 @@ def startup_rendering():
     master.startup_rendering_stack()
 
 
-def start_eyeserver():
+def start_eyeserver(eye_id=None):
     """Starts the pyro eyeserver on the eyes defined by settins.network via fabric"""
+
+    args = argparse.ArgumentParser(description="Starts the pyro5 eye servers on the chosen eyes")
+
+    # adding optional arguments
+    aid = args.add_argument("--eye_id", default=None, help="ID (int) of the eye (nano board) to start eyeserver on as in"
+                                                           "settings.network")
+    args = args.parse_args()
+
+    if args.eye_id is not None:
+        eye_ids = [eye['expected_id'] for eye in network.eyes.values()]
+        eye_id = int(args.eye_id)
+        if eye_id not in eye_ids:
+            raise ValueError(f"Eye ID {eye_id} not found in settings.network")
+        eye_ids = [eye_id]
+    else:
+        eye_ids = [eye['expected_id'] for eye in network.eyes.values()]
+
     logger.info("Starting eye servers...")
     PSWD = getpass('sudo password to start eyeservers: ')
-    eye_ips = [eye['host'] for eye in network.eyes.values()]
+    eye_ips = [eye['host'] for eye in network.eyes.values() if eye['expected_id'] in eye_ids]
     config = Config(overrides={'sudo': {'password': PSWD}})
     eyes = Group(*eye_ips, user=network.nano_username, config=config)
-    for c in eyes:
+    logger.info(f"Starting eyeservers on {eye_ips}")
+    for ci, c in enumerate(eyes):
         c.connect_kwargs.password = PSWD
 
         # checking for already running eyeserver instances
@@ -101,49 +119,52 @@ def start_eyeserver():
             c.run(f'kill -INT -{int(PID)}')
             logger.info(f"Killed eyeserver on host {c.host}, will restart now...")
         else:
-            logger.info(f"Skipping eyeserver startup on host {c.host} as requested...")
+            logger.info(f"Eyeserver stop/restart was not requested on host {c.host}. Maybe no instance was running"
+                        f"or requested to skip restarting.")
 
         # starting a new eye server if it was not running or a restart was requested
         if not found_eye_servers or restart_eyeserver:
             # starting eyeserver
+            logger.info(f'Starting eyeserver on host {c.host}')
             c.run(f'cd {network.nano_cobe_installdir} && '
                   'git pull && '
                   'ls && '
-                  'dtach -n /tmp/tmpdtach '
+                  f'EYE_ID={eye_ids[ci]} dtach -n /tmp/tmpdtach '
                   f'python3 cobe/vision/eye.py --host={c.host} --port={network.unified_eyeserver_port}',
                   hide=True,
                   pty=False)
             logger.info(f'Started eyeserver on host {c.host}')
         time.sleep(5)
 
-        getpass('Eye servers started. Press any key to stop the eye servers...')
+    getpass('Eye servers started. Press any key to stop the eye servers...')
 
-        logger.info('Killing eye-server processes by collected PIDs...')
+    logger.info('Killing eye-server processes by collected PIDs...')
+    for c in eyes:
+        logger.info(f'Stopping eyeserver on host {c.host}')
+        c.connect_kwargs.password = PSWD
+        start_result = c.run('ps ax  | grep "python3 cobe/vision/eye.py"')
+        PID = start_result.stdout.split()[0]  # get PID of first subrocess of python3
+        logger.info(f"Found server process with PID: {PID}, killing it...")
+        # sending INT SIG to the main process will trigger graceful exit (equivalent to KeyboardInterrup)
+        c.run(f'kill -INT -{int(PID)}')
+        logger.info(f"Killed eyeserver on host {c.host}")
+    time.sleep(5)
+
+    # Shutting down the physical nvidia boards if requested
+    is_shutdown = input("Do you want to shutdown the eyes? (y/n): ")
+    if is_shutdown.lower() == "y":
+        logger.info("Shutting down eyes...")
         for c in eyes:
-            logger.info(f'Stopping eyeserver on host {c.host}')
             c.connect_kwargs.password = PSWD
-            start_result = c.run('ps ax  | grep "python3 cobe/vision/eye.py"')
-            PID = start_result.stdout.split()[0]  # get PID of first subrocess of python3
-            logger.info(f"Found server process with PID: {PID}, killing it...")
-            # sending INT SIG to the main process will trigger graceful exit (equivalent to KeyboardInterrup)
-            c.run(f'kill -INT -{int(PID)}')
-            logger.info(f"Killed eyeserver on host {c.host}")
-        time.sleep(5)
+            c.sudo(f'shutdown -h now', warn=True, shell=False)
+            logger.info(f"Shutting down host {c.host}")
+        logger.info("Shutdown complete.")
 
-        # Shutting down the physical nvidia boards if requested
-        is_shutdown = input("Do you want to shutdown the eyes? (y/n): ")
-        if is_shutdown.lower() == "y":
-            logger.info("Shutting down eyes...")
-            for c in eyes:
-                c.connect_kwargs.password = PSWD
-                c.sudo(f'shutdown -h now', warn=True, shell=False)
-                logger.info(f"Shutting down host {c.host}")
-            logger.info("Shutdown complete.")
 
 def stop_eyeserver():
     """Stops the pyro eyeserver on the eyes defined by settins.network via fabric. Can be used when
     eyeserver keeps running in the background blocking new connections"""
-    logger.info("Starting eye servers...")
+    logger.info("Stopping eye servers...")
     PSWD = getpass('sudo password to start eyeservers: ')
     eye_ips = [eye['host'] for eye in network.eyes.values()]
     config = Config(overrides={'sudo': {'password': PSWD}})
@@ -155,13 +176,17 @@ def stop_eyeserver():
             logger.info(f'Stopping eyeserver on host {c.host}')
             c.connect_kwargs.password = PSWD
             start_result = c.run('ps ax  | grep "python3 cobe/vision/eye.py"')
+            num_found_procs = len(start_result.stdout.split("\n"))
             PID = start_result.stdout.split()[0]  # get PID of first subrocess of python3
-            logger.info(f"Found server process with PID: {PID}, killing it...")
-            # sending INT SIG to the main process will trigger graceful exit (equivalent to KeyboardInterrup)
-            c.run(f'kill -INT -{int(PID)}')
-            logger.info(f"Killed eyeserver on host {c.host}")
+            found_eye_servers = num_found_procs > 3
+            if found_eye_servers:
+                logger.info(f"Found server process with PID: {PID}, killing it...")
+                # sending INT SIG to the main process will trigger graceful exit (equivalent to KeyboardInterrup)
+                c.run(f'kill -INT -{int(PID)}')
+                logger.info(f"Killed eyeserver on host {c.host}")
+            else:
+                logger.info(f"Seems like eyeserver is not running on host {c.host}, skipping...")
         time.sleep(5)
-
 
 
 def calibrate(eye_id=-1, on_screen=False):
@@ -172,15 +197,17 @@ def calibrate(eye_id=-1, on_screen=False):
 
     # adding optional arguments
     aid = args.add_argument("--eye_id", default=None, help="ID (int) of the eye (nano board) to be calibrated as in "
-                                                                        "settings.network")
+                                                           "settings.network")
     aos = args.add_argument("--on_screen", default=False, help="If set to True, the calibration target will be "
-                                                                            "displayed on the computer screen instead "
-                                                                            "of the projectors")
+                                                               "displayed on the computer screen instead "
+                                                               "of the projectors")
+    args = args.parse_args()
 
-    if aid is not None:
-        eye_id = aid
-    if aos is not None:
-        on_screen = bool(aos)
+    if args.eye_id is not None:
+        eye_id = int(args.eye_id)
+    if args.on_screen is not None:
+        on_screen = bool(args.on_screen)
+
 
     master = CoBeMaster()
     if on_screen:
